+++ vr/src/main/scala/slinky/vr/AppRegistry.scala
@@ -6,7 +6,7 @@ import scala.scalajs.js
 import scala.scalajs.js.annotation.JSImport
 
 @js.native
-@JSImport("react-vr", "AppRegistry")
+@JSImport("react-360", "AppRegistry")
 object AppRegistry extends js.Object {
   def registerComponent(appKey: String, componentProvider: js.Function0[ReactComponentClass[_]]): Unit = js.native
 }